requirements.txt
discord.py==1.5.1
requests=2.24.0